* show file parameters even if loaded by dblclick on file list item
File parameters (encoding, line ending type and even binary mark)
are shown in a special text field when file is loaded for preview.

It was the case only when user clicks on one of the search results
till now.
Now it's done that way also when user just double-click on file list.